Rewrite without client side auth
Zendesk client now sends a request with the zendesk ticket requester id
to Viper. Viper will then talk back to Zendesk (authenticated from
Viper's end) to Zendesk and ask for the Zendesk user.

Should the request be successful and a user exists in Viper with a
matching email, a user single returned.

If Zendesk says thats user does not exist or Viper cannot find a user
with a matching email, it will return 422 unauthorised.